+++ lib/deimos/active_record_producer.rb
@@ -59,6 +59,29 @@ module Deimos
           k.to_sym != :payload_key && !fields.map(&:name).include?(k)
         end
       end
+
+      # Query to use when polling the database with the DbPoller. Add
+      # includes, joins, or wheres as necessary, or replace entirely.
+      # @param time_from [Time] the time to start the query from.
+      # @param time_to [Time] the time to end the query.
+      # @param column_name [Symbol] the column name to look for.
+      # @param min_id [Numeric] the minimum ID (i.e. all IDs must be greater
+      # than this value).
+      # @return [ActiveRecord::Relation]
+      def poll_query(time_from:, time_to:, column_name: :updated_at, min_id:)
+        klass = config[:record_class]
+        table = ActiveRecord::Base.connection.quote_table_name(klass.table_name)
+        column = ActiveRecord::Base.connection.quote_column_name(column_name)
+        primary = ActiveRecord::Base.connection.quote_column_name(klass.primary_key)
+        klass.where(
+          "((#{table}.#{column} = ? AND #{table}.#{primary} > ?) \
+           OR #{table}.#{column} > ?) AND #{table}.#{column} <= ?",
+          time_from,
+          min_id,
+          time_from,
+          time_to
+        )
+      end
     end
   end
 end