+++ lib/deimos/utils/db_producer.rb
@@ -2,7 +2,8 @@
 
 module Deimos
   module Utils
-    # Class which continually polls the database and sends Kafka messages.
+    # Class which continually polls the kafka_messages table
+    # in the database and sends Kafka messages.
     class DbProducer
       include Phobos::Producer
       attr_accessor :id, :current_topic